+++ server/schemas/Condiment.js
@@ -1,9 +1,9 @@
-const { Text, Decimal } = require("@keystonejs/fields");
+const { Text, Float } = require("@keystonejs/fields");
 
 module.exports = {
   fields: {
     name: { type: Text },
-    price: { type: Decimal },
-    stock: { type: Decimal },
+    price: { type: Float },
+    stock: { type: Float },
   },
 };
